feat: 백엔드 연동을 대비한 api 분리
